Added in a default script and call timing support

+++ playpen/v2_plugins/harness/harness.py
@@ -13,6 +13,7 @@
 
 import base64
 import ConfigParser
+import datetime
 import httplib
 import json
 import optparse
@@ -25,6 +26,7 @@ import prompt
 
 REST_COLOR = prompt.COLOR_LIGHT_BLUE
 RESPONSE_COLOR = prompt.COLOR_LIGHT_PURPLE
+TIME_COLOR = prompt.COLOR_LIGHT_RED
 
 # -- classes ------------------------------------------------------------------
 
@@ -51,19 +53,19 @@ class Harness:
         Runs the appropriate commands accroding the script given at instantiation.
         """
 
-        if self.script.getboolean('general', 'run_delete_repo'):
+        if self.script.has_option('general', 'run_delete_repo') and self.script.getboolean('general', 'run_delete_repo'):
             self.delete_repo()
 
-        if self.script.getboolean('general', 'run_create_repo'):
+        if self.script.has_option('general', 'run_create_repo') and self.script.getboolean('general', 'run_create_repo'):
             self.create_repo()
 
-        if self.script.getboolean('general', 'run_add_importer'):
+        if self.script.has_option('general', 'run_add_importer') and self.script.getboolean('general', 'run_add_importer'):
             self.add_importer()
 
-        if self.script.getboolean('general', 'run_sync_repo'):
+        if self.script.has_option('general', 'run_sync_repo') and self.script.getboolean('general', 'run_sync_repo'):
             self.sync_repo()
 
-        if self.script.getboolean('general', 'run_list_units'):
+        if self.script.has_option('general', 'run_list_units') and self.script.getboolean('general', 'run_list_units'):
             self.list_units()
 
     def delete_repo(self):
@@ -73,14 +75,14 @@ class Harness:
         url = '/v2/repositories/%s/' % repo_id
 
         self.prompt.write('Looking up repository [%s] on the server' % repo_id)
-        self._print_rest('GET', url, None)
+        self._call('GET', url, None)
         self._pause()
 
         status, body = self.connection.GET(url)
 
         if status == 200:
             self.prompt.write('Repository with ID [%s] already exists and will be deleted' % repo_id)
-            self._print_rest('DELETE', url, None)
+            self._call('DELETE', url, None)
             self._pause()
             self.connection.DELETE(url)
         elif status == 404:
@@ -99,7 +101,7 @@ class Harness:
         }
 
         self.prompt.write('Creating repository with ID [%s]' % repo_id)
-        self._print_rest('POST', url, body)
+        self._call('POST', url, body)
         self._pause()
 
         status, body = self.connection.POST(url, body)
@@ -129,7 +131,7 @@ class Harness:
         for k, v in importer_config.items():
             self.prompt.write('    %-15s : %s' % (k, v))
 
-        self._print_rest('POST', url, body)
+        self._call('POST', url, body)
         self._pause()
         
         status, body = self.connection.POST(url, body)
@@ -154,7 +156,7 @@ class Harness:
         }
 
         self.prompt.write('Synchronizing repository [%s]' % repo_id)
-        self._print_rest('POST', url, body)
+        self._call('POST', url, body)
         self._pause()
 
         status, body = self.connection.POST(url, body)
@@ -163,6 +165,9 @@ class Harness:
 
         self.prompt.write('Synchronization complete')
 
+    def load_sync_history(self):
+        pass
+
     def list_units(self):
         self._print_divider()
 
@@ -170,7 +175,7 @@ class Harness:
         url = '/v2/content/%s/units/' % type_id
 
         self.prompt.write('Retrieving list of units of type [%s]' % type_id)
-        self._print_rest('GET', url, None)
+        self._call('GET', url, None)
         self._pause()
 
         status, body = self.connection.GET(url)
@@ -179,6 +184,7 @@ class Harness:
         self._print_response(status, body, limit=unit_limit)
 
         self.prompt.write('Retrieved [%d] units' % len(body))
+
 
     # -- utilities ------------------------------------------------------------
 
@@ -196,18 +202,36 @@ class Harness:
         self.prompt.write('')
         
     def _print_divider(self):
+        self.prompt.write('')
         self.prompt.write('=' * 60)
 
-    def _print_rest(self, method, url, body):
-        if not self.script.getboolean('output', 'print_rest'):
-            return
+    def _call(self, method, url, body):
+        print_rest = self.script.getboolean('output', 'print_rest')
+        print_call_time = self.script.getboolean('output', 'print_call_time')
 
-        self.prompt.write('')
-        self.prompt.write(self.prompt.color('Method: %s' % method, REST_COLOR))
-        self.prompt.write(self.prompt.color('URL:    %s' % url, REST_COLOR))
-        if body is not None:
-            self.prompt.write(self.prompt.color('Body:   %s' % body, REST_COLOR))
-        self.prompt.write('')
+        start = datetime.datetime.now()
+        func = getattr(self.connection, method)
+        if body is None:
+            r_status, r_body = func(url)
+        else:
+            r_status, r_body = func(url, body)
+        end = datetime.datetime.now()
+
+        if print_rest:
+            self.prompt.write('')
+            self.prompt.write(self.prompt.color('Method: %s' % method, REST_COLOR))
+            self.prompt.write(self.prompt.color('URL:    %s' % url, REST_COLOR))
+            if body is not None:
+                self.prompt.write(self.prompt.color('Body:' % body, REST_COLOR))
+                indent = self.script.getint('output', 'json_indent')
+                formatted_body = json.dumps(body, indent=indent)
+                self.prompt.write(self.prompt.color(formatted_body, REST_COLOR))
+
+        if print_call_time:
+            ellapsed = end - start
+            self.prompt.write(self.prompt.color('Call Time: %d seconds' % ellapsed.seconds, TIME_COLOR))
+
+        return r_status, r_body
 
     def _print_response(self, status, body, limit=None):
         if not self.script.getboolean('output', 'print_rest'):
@@ -271,7 +295,10 @@ class PulpConnection:
                                 headers=auth_header)
         response = self.connection.getresponse()
         response_body = response.read()
-        response_body = json.loads(response_body)
+        try:
+            response_body = json.loads(response_body)
+        except:
+            pass
 
         return response.status, response_body
 
@@ -294,6 +321,7 @@ if __name__ == '__main__':
         sys.exit(1)
 
     script = ConfigParser.SafeConfigParser()
+    script.read('scripts/_default.ini')
     script.read(options.script)
 
     connection = PulpConnection()
